requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt
#
apscheduler==3.6.3
    # via python-telegram-bot
certifi==2020.12.5
    # via
    #   python-telegram-bot
    #   requests
cffi==1.14.4
    # via cryptography
chardet==4.0.0
    # via requests
cryptography==3.3.1
    # via python-telegram-bot
decorator==4.4.2
    # via python-telegram-bot
idna==2.10
    # via requests
pycparser==2.20
    # via cffi
pymongo==3.11.2
    # via rysolv_monitor (setup.py)
python-telegram-bot==13.1
    # via rysolv_monitor (setup.py)
pytz==2020.5
    # via
    #   apscheduler
    #   python-telegram-bot
    #   tzlocal
requests==2.25.1
    # via rysolv_monitor (setup.py)
six==1.15.0
    # via
    #   apscheduler
    #   cryptography
tornado==6.1
    # via python-telegram-bot
tzlocal==2.1
    # via apscheduler
urllib3==1.26.2
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
